pyproject.toml
[tool.poetry]
name = "Pbot"
version = "0.2.0"
description = "Pbot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.1.0"
gino = "^1.0.1"
alembic = "^1.4.3"
psycopg2 = "^2.8.6"
feedparser = "^6.0.2"
bs4 = "^0.0.1"
lxml = "^4.6.1"
langdetect = "^1.0.8"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
